Refactor SegmentLoadDropHandler code.
