+++ server/src/main/java/org/apache/druid/server/coordination/ChangeRequestsSnapshot.java
@@ -24,7 +24,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.Preconditions;
 
 import javax.annotation.Nullable;
-
 import java.util.List;
 
 /**
