+++ server/src/test/java/org/apache/druid/server/coordination/ZkCoordinatorTest.java
@@ -23,7 +23,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableMap;
 import org.apache.curator.utils.ZKPaths;
 import org.apache.druid.curator.CuratorTestBase;
-import org.apache.druid.guice.ServerTypeConfig;
 import org.apache.druid.java.util.common.Intervals;
 import org.apache.druid.java.util.emitter.EmittingLogger;
 import org.apache.druid.segment.IndexIO;
@@ -103,12 +102,8 @@ public class ZkCoordinatorTest extends CuratorTestBase
     SegmentLoadDropHandler segmentLoadDropHandler = new SegmentLoadDropHandler(
         new SegmentLoaderConfig(),
         EasyMock.createNiceMock(DataSegmentAnnouncer.class),
-        EasyMock.createNiceMock(DataSegmentServerAnnouncer.class),
         EasyMock.createNiceMock(SegmentManager.class),
-        EasyMock.createNiceMock(ScheduledExecutorService.class),
-        new ServerTypeConfig(ServerType.HISTORICAL),
-        new TestCoordinatorClient(),
-        new NoopServiceEmitter()
+        EasyMock.createNiceMock(ScheduledExecutorService.class)
     )
     {
       @Override
